mudando as coisas pro jeito do severo
